[REF] Merge Shell and Reading files into Helpers
